requirements.txt
allure-pytest==2.13.2
python-dotenv==1.0.0
httpx==0.25.0
jsonpath-ng==1.6.0
pydantic==2.3.0
pytest==7.4.2
pytest-xdist==3.3.1